Move files into new domain and add Command for cleanup
