+++ app/Http/Controllers/PasswordResetController.php
@@ -4,11 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Entities\Models\Eloquent\AccountPasswordReset;
 use App\Exceptions\Http\NotFoundException;
-use App\Http\Actions\AccountPasswordReset\ResetAccountPassword;
-use App\Http\Actions\AccountPasswordReset\SendPasswordResetEmail;
 use App\Http\Requests\ResetPasswordRequest;
 use App\Http\Requests\SendPasswordEmailRequest;
 use App\Http\WebController;
+use Domain\PasswordReset\UseCases\ResetAccountPassword;
+use Domain\PasswordReset\UseCases\SendPasswordResetEmail;
 use Illuminate\Http\Request;
 
 final class PasswordResetController extends WebController
